♻️ refactor: rename folder (tests/flet) to (tests/flet-app) and update to v0.2.4
