+++ tests/flet-app/assets/icon.png
