+++ tests/flet-app/controllers/counter.py
@@ -1,4 +1,5 @@
 import flet as ft
+
 import flet_easy as fs
