+++ tests/flet-app/main.py
@@ -1,13 +1,15 @@
 from pathlib import Path
 
-import flet_easy as fs
 from core.config import ConfigApp
 
 # from core.sensitive import PRIVATE_KEY, PUBLIC_KEY
 from core.sensitive import SECRET_KEY
 from models import create_tables
 
+import flet_easy as fs
+
 create_tables()
+
 app = fs.FletEasy(
     route_init="/home",
     route_login="/login",
@@ -19,11 +21,14 @@ app = fs.FletEasy(
 
 
 def starting_page(data: fs.Datasy):
-    print(f"\n⚡[MIDLEWARE RUNNING]\n-> Route:{data.route}\n1. Loading the page")
+    """Remove [print()] from Python if build Windows"""
+    print(f"\n⚡[MIDDLEWARES RUNNING]\n-> Route:{data.route}\n1. Loading the page")
 
 
 async def starting_page_two(data: fs.Datasy):
+    """Remove [print()] from Python if build Windows"""
     print("2. Loading the page")
+
     # We allow the following paths, otherwise it redirects to the path ('/login').
     # Note: To check the urls you should use (data.route) and not (data.page.route), as this will cause an infinite loop.
     routes = [
@@ -41,9 +46,15 @@ async def starting_page_two(data: fs.Datasy):
         return data.redirect("/login")
 
 
+""" [If you have any problem - build Windows]
+Remove [print()] from Python if used in 'add_middleware' functions. """
 app.add_middleware([starting_page, starting_page_two])
+
 ConfigApp(app)
 
-# We run the application
+# We run the application web
 # run =app.run(export_asgi_app=True)
+
+""" [If you have any problem - build web]
+Use ft.app(target=app.run(fastapi=true)) when compiling web statica. """
 app.run()
