+++ tests/flet-app/pyproject.toml
@@ -1,8 +1,8 @@
 [project]
-name = "flet"
+name = "flet-app"
 version = "0.1.0"
 license = "MIT"
-dependencies = ["flet==0.22.*", "flet-easy==0.2.*", "peewee"]
+dependencies = ["flet==0.24.*", "flet-easy[jwt]==0.2.*", "peewee"]
 
 [[project.authors]]
 name = "Daxexs"
